requirements.txt
# Core RAG System Dependencies

# Web Framework
Flask
werkzeug
requests 

# AI/ML - Generation
google-genai
groq
mistralai
cerebras-cloud-sdk

# AI/ML - Embeddings (embedserver.py)
# onnxruntime-gpu        
transformers           
numpy                  

# Document Processing
PyMuPDF                 
python-docx              
python-pptx         
openpyxl                
Pillow               

# Vector Database
chromadb

# Utilities
python-dotenv

# Database
psycopg2-binary         

# Authentication
PyJWT                  # JWT token handling
bcrypt                 # Password hashing

# Enhanced Retrieval (Vector Search + Reranking)
rank-bm25               # BM25 search for hybrid retrieval
nltk                    # Sentence tokenization for smart chunking
tiktoken                # Token counting for context limit

# S3/R2 Storage
boto3

# Image Generation
fake-useragent          # User agent for API requests
httpx                   # HTTP client for async requests